Bump up furiosa-smi to latest

+++ pkg/smi/binding/zz_types.go
@@ -10,25 +10,25 @@ package binding
 */
 import "C"
 
-// FuriosaSmiDeviceHandle type as declared in smi/furiosa_smi.h:58
+// FuriosaSmiDeviceHandle type as declared in smi/furiosa_smi.h:62
 type FuriosaSmiDeviceHandle uint32
 
-// FuriosaSmiDeviceHandles as declared in smi/furiosa_smi.h:63
+// FuriosaSmiDeviceHandles as declared in smi/furiosa_smi.h:67
 type FuriosaSmiDeviceHandles struct {
-	Count          uint32
-	DeviceHandles  [64]FuriosaSmiDeviceHandle
+	Count         uint32
+	DeviceHandles [64]FuriosaSmiDeviceHandle
 }
 
-// FuriosaSmiVersion as declared in smi/furiosa_smi.h:71
+// FuriosaSmiVersion as declared in smi/furiosa_smi.h:75
 type FuriosaSmiVersion struct {
-	Arch           FuriosaSmiArch
-	Major          uint32
-	Minor          uint32
-	Patch          uint32
-	Metadata       [96]byte
+	Arch     FuriosaSmiArch
+	Major    uint32
+	Minor    uint32
+	Patch    uint32
+	Metadata [96]byte
 }
 
-// FuriosaSmiDeviceInfo as declared in smi/furiosa_smi.h:85
+// FuriosaSmiDeviceInfo as declared in smi/furiosa_smi.h:89
 type FuriosaSmiDeviceInfo struct {
 	Arch            FuriosaSmiArch
 	CoreNum         uint32
@@ -43,26 +43,26 @@ type FuriosaSmiDeviceInfo struct {
 	DriverVersion   FuriosaSmiVersion
 }
 
-// FuriosaSmiDeviceFile as declared in smi/furiosa_smi.h:91
+// FuriosaSmiDeviceFile as declared in smi/furiosa_smi.h:95
 type FuriosaSmiDeviceFile struct {
-	CoreStart      uint32
-	CoreEnd        uint32
-	Path           [256]byte
+	CoreStart uint32
+	CoreEnd   uint32
+	Path      [256]byte
 }
 
-// FuriosaSmiDeviceFiles as declared in smi/furiosa_smi.h:96
+// FuriosaSmiDeviceFiles as declared in smi/furiosa_smi.h:100
 type FuriosaSmiDeviceFiles struct {
-	Count          uint32
-	DeviceFiles    [64]FuriosaSmiDeviceFile
+	Count       uint32
+	DeviceFiles [64]FuriosaSmiDeviceFile
 }
 
-// FuriosaSmiCoreStatuses as declared in smi/furiosa_smi.h:101
+// FuriosaSmiCoreStatuses as declared in smi/furiosa_smi.h:105
 type FuriosaSmiCoreStatuses struct {
-	Count          uint32
-	CoreStatus     [128]FuriosaSmiCoreStatus
+	Count      uint32
+	CoreStatus [128]FuriosaSmiCoreStatus
 }
 
-// FuriosaSmiDeviceErrorInfo as declared in smi/furiosa_smi.h:113
+// FuriosaSmiDeviceErrorInfo as declared in smi/furiosa_smi.h:117
 type FuriosaSmiDeviceErrorInfo struct {
 	AxiPostErrorCount      uint32
 	AxiFetchErrorCount     uint32
@@ -75,40 +75,40 @@ type FuriosaSmiDeviceErrorInfo struct {
 	DeviceErrorCount       uint32
 }
 
-// FuriosaSmiDriverInfo as declared in smi/furiosa_smi.h:118
+// FuriosaSmiDriverInfo as declared in smi/furiosa_smi.h:122
 type FuriosaSmiDriverInfo struct {
-	Count          uint32
-	DriverInfo     [24]FuriosaSmiVersion
+	Count      uint32
+	DriverInfo [24]FuriosaSmiVersion
 }
 
-// FuriosaSmiPeUtilization as declared in smi/furiosa_smi.h:125
+// FuriosaSmiPeUtilization as declared in smi/furiosa_smi.h:129
 type FuriosaSmiPeUtilization struct {
 	CoreCount         uint32
 	Cores             [64]uint32
 	TimeWindowMil     uint32
-	PeUsagePercentage uint32
+	PeUsagePercentage float64
 }
 
-// FuriosaSmiMemoryUtilization as declared in smi/furiosa_smi.h:130
+// FuriosaSmiMemoryUtilization as declared in smi/furiosa_smi.h:134
 type FuriosaSmiMemoryUtilization struct {
-	TotalBytes     uint64
-	InUseBytes     uint64
+	TotalBytes uint64
+	InUseBytes uint64
 }
 
-// FuriosaSmiDeviceUtilization as declared in smi/furiosa_smi.h:136
+// FuriosaSmiDeviceUtilization as declared in smi/furiosa_smi.h:140
 type FuriosaSmiDeviceUtilization struct {
-	PeCount        uint32
-	Pe             [64]FuriosaSmiPeUtilization
-	Memory         FuriosaSmiMemoryUtilization
+	PeCount uint32
+	Pe      [64]FuriosaSmiPeUtilization
+	Memory  FuriosaSmiMemoryUtilization
 }
 
-// FuriosaSmiDevicePowerConsumption as declared in smi/furiosa_smi.h:140
+// FuriosaSmiDevicePowerConsumption as declared in smi/furiosa_smi.h:144
 type FuriosaSmiDevicePowerConsumption struct {
-	RmsTotal       float64
+	RmsTotal float64
 }
 
-// FuriosaSmiDeviceTemperature as declared in smi/furiosa_smi.h:145
+// FuriosaSmiDeviceTemperature as declared in smi/furiosa_smi.h:149
 type FuriosaSmiDeviceTemperature struct {
-	SocPeak        int32
-	Ambient        int32
+	SocPeak float64
+	Ambient float64
 }
